+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+2.0.1
+=====
+
+*   (bug) Properly write `_dependencies.json` if anything changes.
+*   (bug) Remove old builds in SCSS watcher.
+
+
 2.0.0
 =====